+++ lib/completion_kit/engine.rb
@@ -5,12 +5,6 @@ module CompletionKit
   class Engine < ::Rails::Engine
     isolate_namespace CompletionKit
 
-    initializer("completion_kit.inflections", before: :load_config_initializers) do
-      ActiveSupport::Inflector.inflections(:en) do |inflect|
-        inflect.irregular "criterion", "criteria"
-      end
-    end
-
     paths.add "app/services", eager_load: true
 
     def self.register_assets(app)